+++ CHANGELOG.md
@@ -33,6 +33,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Recent called list is niet meer cursed AF
 - Timer van phone notifications die optelde werken nu terug treffelijk
 - Items resetten nu terug naar originele rotatie wanneer je ze loslaat buiten de toegestane grid
+- Je kan weer flyers aanvragen
 
 ## [1.4.2]